Center images on recipes, slight styling changes
Created a new "images" class to separate styling from the index.html
images using "center" class. This allows the images to be centered and
use the original image aspect ratio. Some images are shaped oddly and
may need to be swapped

Readjusted color of header 1

Still need to reformat the links on recipes page as last step.
- Back to home button
- Other Recipes section will be removed, and formatted next to current
recipe's image. May potentially use faded images to click, or just words
on left and right